ai-pipeline/requirements.txt
# AI Pipeline Requirements for Feedback Analyzer
# Generated: 2026-03-16

# Core
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
pandas>=2.0.0

# Preprocessing
spacy>=3.6.0
nltk>=3.8.0
pyspellchecker>=0.7.0
huggingface-hub>=0.19.0
pyxdameraulevenshtein>=1.7.0

# Topic Modeling
# Topic modeling deprecated: heavy local deps were removed in favor of MCP-backed ABSA
# (bertopic, sentence-transformers, umap-learn, hdbscan removed)

# HTTP client for MCP
requests>=2.28.0

# Classification (BERT/RoBERTa)
accelerate>=0.20.0

# Database
supabase>=2.0.0
python-dotenv>=1.0.0

# Web Server
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Utilities
tqdm>=4.65.0
scikit-learn>=1.3.0
joblib>=1.3.0
ipython>=8.0.0

# Development/Testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
